% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a}
\alias{a}
\title{Example data for HW 2}
\format{An object of class \code{matrix} with 3 rows and 3 columns.}
\usage{
a
}
\description{
Example data for HW 2
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q2p3.rda}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
