% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleanData}
\alias{cleanData}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 150 rows and 5 columns.}
\usage{
cleanData
}
\description{
This is data to be included in my package
}
\references{
\url{https://archive.ics.uci.edu/ml/datasets/iris}
}
\author{
Huangying Kuang \email{kuang048@umn.edu}
}
\keyword{datasets}
