% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW2Functions.R
\name{\%question2\%}
\alias{\%question2\%}
\title{Given a numeric matrix A and a numeric vector x, calculates xTA−1x (binary operator VERSION)}
\usage{
A \%question2\% x
}
\arguments{
\item{A}{matrix}

\item{x}{vector}
}
\value{
value of xTA−1x
}
\description{
Given a numeric matrix A and a numeric vector x, calculates xTA−1x (binary operator VERSION)
}
\examples{
matrix(rnorm(20),nc=5,nr=4) \%question2\% c(7,8,9)
}
