% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW2Functions.R
\name{hw2_1}
\alias{hw2_1}
\title{Given a numeric matrix A and a numeric vector x, calculates xTA−1x}
\usage{
hw2_1(A, x)
}
\arguments{
\item{A}{matrix}

\item{x}{vector}
}
\value{
value of xTA−1x
}
\description{
Given a numeric matrix A and a numeric vector x, calculates xTA−1x
}
\examples{
hw2_1(matrix(rnorm(20),nc=5,nr=4),c(7,8,9))
}
