% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW2Functions.R
\name{myapply}
\alias{myapply}
\title{A function similar to apply()}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{matrix of any type}

\item{MARGIN}{either (the number) 1 or (the number) 2}

\item{FUN}{R function that maps vectors to vectors}

\item{...}{optional parameters for plotting}
}
\value{
list
}
\description{
A function similar to apply()
}
\examples{
myapply(matrix(1:6, nrow = 3, ncol = 2),1,mean)
}
