% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW2Functions.R
\name{q3}
\alias{q3}
\title{Column standardization}
\usage{
q3(m)
}
\arguments{
\item{m}{matrix}
}
\value{
standardized matrix
}
\description{
Takes a numeric matrix and standardizes its columns
}
\examples{
q3(matrix(rnorm(20),nc=5,nr=4))
}
