% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW2Functions.R
\name{q4}
\alias{q4}
\title{Column standardization (not using loop)}
\usage{
q4(m)
}
\arguments{
\item{m}{matrix}
}
\value{
standardized matrix
}
\description{
Takes a numeric matrix and standardizes its columns
}
\examples{
q4(matrix(rnorm(20),nc=5,nr=4))
}
