% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW1Functions.R, R/func7.R
\name{func7}
\alias{func7}
\alias{func7}
\title{MLE}
\usage{
func7(x, fun, interval)

func7(x, fun, interval)
}
\arguments{
\item{x}{vector}

\item{fun}{a distribution function}

\item{interval}{vector, i.e., interval for optimize function}

\item{func}{function, e.g., `function(theta, x) dgamma(x, shape = theta, log = TRUE)`}

\item{x}{dataset}

\item{interval}{the interval where the optimization shoul search among}
}
\value{
scalar

return the log of the ppdf or pmf
}
\description{
Computes the liklihood of a given distribution for data x

the function from homework 1, problem 7
}
\examples{
func1 = function(theta, x) dgamma(x, shape = theta, log = TRUE)
result7_gamma <- func7(x1,func1,c(0,3))
result7_gamma

x <- scan(url("http://www.stat.umn.edu/geyer/3701/data/q1p3.txt"))
f <- function(theta, x) dgamma(x, shape = theta, log = TRUE)
interval <- c(0,100)
func7(x,f,interval)

}
