% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func7.R
\name{func7}
\alias{func7}
\title{the function from homework 1, problem 7}
\usage{
func7(x, fun, interval)
}
\arguments{
\item{x}{dataset}

\item{fun}{a distribution function}

\item{interval}{the interval where the optimization shoul search among}
}
\value{
return the log of the ppdf or pmf
}
\description{
the function from homework 1, problem 7
}
\examples{
x <- scan(url("http://www.stat.umn.edu/geyer/3701/data/q1p3.txt"))
f <- function(theta, x) dgamma(x, shape = theta, log = TRUE)
interval <- c(0,100)
func7(x,f,interval)

}
